# komplett-server-dev/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node-Typescript API",
    "version": "1.0.0",
    "description": "A sample API"
  },
  "servers": [
    {
      "url": "http://192.168.0.119:3000"
    },
    {
      "url": "https://mongo-kompletee.herokuapp.com"
    },
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://ec2-18-216-130-99.us-east-2.compute.amazonaws.com:3000"
    },
    {
      "url": "http://ec2-3-22-60-204.us-east-2.compute.amazonaws.com:3000"
    }
  ],
  "paths": {
...
    "/api/customer": {
      "post": {
        "description": "create customer",
        "tags": [
          "customer"
        ],
        "requestBody": {
          "description": "customer creation request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerSchema"
              },
              "example": {
                "user": {
                  "name": "someone",
                  "email": "someemail@some.com"
                },
                "contactPerson": "Mohcine",
                "location": {
                  "city": "oslo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created customer",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CustomerSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Get Customers api with pagination and sorting",
        "tags": [
          "customer"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = createdAt"
          },
          {
            "in": "query",
            "name": "customerName",
            "schema": {
              "type": "string"
            },
            "description": "Example = somename"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "acs"
            },
            "description": "Example = acs"
          }
        ],
        "responses": {
          "200": {
            "description": "returns customers list",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CustomerSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "update customer",
        "tags": [
          "customer"
        ],
        "requestBody": {
          "description": "_id field is required",
          "required": true,
          "content": {
            "application/json": {
              "required": [
                "id"
              ],
              "schema": {
                "$ref": "#/components/schemas/CustomerSchema"
              },
              "example": {
                "_id": "5f9bbb17405a9420e895920f",
                "user": {
                  "_id": "5f9bbb17405a9420e895920d",
                  "name": "Mohcine",
                  "email": "someemail@some.com"
                },
                "contactPerson": "Mohcine",
                "location": {
                  "_id": "5f9bbb17405a9420e895920e",
                  "city": "oslo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created customer",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CustomerSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/customer/{id}": {
      "get": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/CustomerSchema"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete customer by id",
        "tags": [
          "customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
    "/api/employee": {
      "get": {
        "description": "Get Employees api with pagination and sorting",
        "tags": [
          "employee"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = createdAt"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "acs"
            },
            "description": "Example = acs"
          }
        ],
        "responses": {
          "200": {
            "description": "returns customers list",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/EmployeeSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "create employee",
        "tags": [
          "employee"
        ],
        "requestBody": {
          "description": "employee creation request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeSchema"
              },
              "example": {
                "user": {
                  "name": "someone",
                  "email": "someemail@some.com"
                },
                "hseCardNumber": 5,
                "employeeNumber": 1,
                "location": {
                  "city": "oslo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created employee",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/EmployeeSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "update employee",
        "tags": [
          "employee"
        ],
        "requestBody": {
          "description": "_id field is required",
          "required": true,
          "content": {
            "application/json": {
              "required": [
                "id"
              ],
              "schema": {
                "$ref": "#/components/schemas/EmployeeSchema"
              },
              "example": {
                "_id": "5f9bbb17405a9420e895920f",
                "user": {
                  "_id": "5f9bbb17405a9420e895920d",
                  "name": "Mohcine",
                  "email": "someemail@some.com"
                },
                "hseCardNumber": 15,
                "location": {
                  "_id": "5f9bbb17405a9420e895920e",
                  "city": "oslo",
                  "postCode": 548
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return created employee",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/EmployeeSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/employee/{id}": {
      "get": {
        "description": "get submaterials for the given materialId",
        "tags": [
          "employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/EmployeeSchema"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete employee by id",
        "tags": [
          "employee"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
...
    "/api/project/customer/{id}": {
      "get": {
        "description": "get project by customer id",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": {
                    "$ref": "#/components/schemas/ProjectSchema"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomerSchema": {
        "properties": {
          "userId": {
            "type": "string"
          },
          "contactPerson": {
            "type": "string"
          },
          "customerType": {
            "type": "string"
          },
          "contactPersonNumber": {
            "type": "string"
          },
          "organizationNumber": {
            "type": "string"
          },
          "contactPersonEmail": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "$ref": "#/components/schemas/LocationSchema"
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          }
        }
      },
      "EmployeeSchema": {
        "properties": {
          "employeeNumber": {
            "type": "number"
          },
          "hseCardNumber": {
            "type": "number"
          },
          "otherContactDetails": {
            "type": "string"
          },
          "emergencyContactName": {
            "type": "string"
          },
          "emergencyContactPhoneNumber": {
            "type": "string"
          },
          "emergencyContactEmail": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "$ref": "#/components/schemas/LocationSchema"
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          }
        }
      },
      "LocationSchema": {
        "properties": {
          "name": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "stateCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "postCode": {
            "type": "string"
          },
          "formattedAddress": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        }
      },
...
      "ProjectSchema": {
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "contactPersonName": {
            "type": "string"
          },
          "contactPersonEmail": {
            "type": "string"
          },
          "contactPersonPhoneNumber": {
            "type": "string"
          },
          "projectNumber": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "estimatedHours": {
            "type": "number"
          },
          "registeredHours": {
            "type": "number"
          },
          "startDate": {
            "type": "date"
          },
          "completionDate": {
            "type": "date"
          },
          "deadLine": {
            "type": "date"
          },
          "locationId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "$ref": "#/components/schemas/LocationSchema"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSchema"
          }
        }
      },
...
      "UserSchema": {
        "required": [
          "email",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "passwordResetToken": {
            "type": "string"
          },
          "passwordResetExpires": {
            "type": "string",
            "format": "date"
          },
          "tokens": {
            "type": "array"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Error description",
            "example": "User created"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
...
}